Move GPDB related API from tuplesort.h to tuplesort_gp.h
This includes functions related to mk sort & reader/writer
