+++ src/backend/optimizer/path/costsize.c
@@ -76,6 +76,7 @@
 #include "utils/lsyscache.h"
 #include "utils/selfuncs.h"
 #include "utils/tuplesort.h"
+#include "utils/tuplesort_gp.h"
 
 #include "cdb/cdbpath.h"        /* cdbpath_rows() */
 #include "cdb/cdbvars.h"
