+++ src/backend/utils/sort/tuplesort.c
@@ -117,6 +117,7 @@
 #include "utils/pg_rusage.h"
 #include "utils/syscache.h"
 #include "utils/tuplesort.h"
+#include "utils/tuplesort_gp.h"
 
 #include "cdb/cdbvars.h"
