+++ src/backend/utils/sort/tuplesort_mkqsort.c
@@ -13,6 +13,7 @@
 #include "postgres.h"
 #include "access/genam.h"
 #include "utils/tuplesort.h"
+#include "utils/tuplesort_gp.h"
 #include "utils/tuplesort_mk.h"
 
 #include "miscadmin.h"
