+++ src/include/utils/tuplesort.h
@@ -36,9 +36,6 @@ struct StringInfoData;                  /* #include "lib/stringinfo.h" */
  */
 typedef struct TuplesortPos TuplesortPos;
 typedef struct Tuplesortstate Tuplesortstate;
-typedef struct TuplesortPos_mk TuplesortPos_mk;
-typedef struct Tuplesortstate_mk Tuplesortstate_mk;
-
 /*
  * We provide two different interfaces to what is essentially the same
  * code: one for sorting HeapTuples and one for sorting IndexTuples.
@@ -54,14 +51,8 @@ typedef struct Tuplesortstate_mk Tuplesortstate_mk;
  * Yet a third slightly different interface supports sorting bare Datums.
  */
 extern void tuplesort_begin_pos(Tuplesortstate *state, TuplesortPos **pos);
-extern void tuplesort_begin_pos_mk(Tuplesortstate_mk *state, TuplesortPos_mk **pos);
 
 extern Tuplesortstate *tuplesort_begin_heap(TupleDesc tupDesc,
-					 int nkeys, AttrNumber *attNums,
-					 Oid *sortOperators, bool *nullsFirstFlags,
-					 int workMem, bool randomAccess);
-extern Tuplesortstate_mk *tuplesort_begin_heap_mk(ScanState * ss,
-					 TupleDesc tupDesc,
 					 int nkeys, AttrNumber *attNums,
 					 Oid *sortOperators, bool *nullsFirstFlags,
 					 int workMem, bool randomAccess);
@@ -73,70 +64,41 @@ extern Tuplesortstate *tuplesort_begin_heap_file_readerwriter(
 		int nkeys, AttrNumber *attNums,
 		Oid *sortOperators, bool *nullsFirstFlags,
 		int workMem, bool randomAccess);
-extern Tuplesortstate_mk *tuplesort_begin_heap_file_readerwriter_mk(
-		ScanState * ss,
-		const char* rwfile_prefix, bool isWriter,
-		TupleDesc tupDesc, 
-		int nkeys, AttrNumber *attNums,
-		Oid *sortOperators, bool *nullsFirstFlags,
-		int workMem, bool randomAccess);
 
 extern Tuplesortstate *tuplesort_begin_index(Relation indexRel,
-					  bool enforceUnique,
-					  int workMem, bool randomAccess);
-extern Tuplesortstate_mk *tuplesort_begin_index_mk(Relation indexRel,
 					  bool enforceUnique,
 					  int workMem, bool randomAccess);
 
 extern Tuplesortstate *tuplesort_begin_datum(Oid datumType,
 					  Oid sortOperator, bool nullsFirstFlag,
 					  int workMem, bool randomAccess);
-extern Tuplesortstate_mk *tuplesort_begin_datum_mk(ScanState * ss,
-					  Oid datumType,
-					  Oid sortOperator, bool nullsFirstFlag,
-					  int workMem, bool randomAccess);
 
 extern void cdb_tuplesort_init(Tuplesortstate *state, int unique,
 							   int sort_flags,
 							   int64 maxdistinct);
-extern void cdb_tuplesort_init_mk(Tuplesortstate_mk *state, int unique,
-							   int sort_flags,
-							   int64 maxdistinct);
 
 extern void tuplesort_set_bound(Tuplesortstate *state, int64 bound);
-extern void tuplesort_set_bound_mk(Tuplesortstate_mk *state, int64 bound);
 
 extern void tuplesort_puttupleslot(Tuplesortstate *state, TupleTableSlot *slot);
-extern void tuplesort_puttupleslot_mk(Tuplesortstate_mk *state, TupleTableSlot *slot);
 
 extern void tuplesort_putindextuple(Tuplesortstate *state, IndexTuple tuple);
-extern void tuplesort_putindextuple_mk(Tuplesortstate_mk *state, IndexTuple tuple);
 
 extern void tuplesort_putdatum(Tuplesortstate *state, Datum val, bool isNull);
-extern void tuplesort_putdatum_mk(Tuplesortstate_mk *state, Datum val, bool isNull);
 
 extern void tuplesort_performsort(Tuplesortstate *state);
-extern void tuplesort_performsort_mk(Tuplesortstate_mk *state);
 
 extern bool tuplesort_gettupleslot_pos(Tuplesortstate *state, TuplesortPos *pos, bool forward, TupleTableSlot *slot, MemoryContext mcontext);
-extern bool tuplesort_gettupleslot_pos_mk(Tuplesortstate_mk *state, TuplesortPos_mk *pos, bool forward, TupleTableSlot *slot, MemoryContext mcontext);
 
 extern bool tuplesort_gettupleslot(Tuplesortstate *state, bool forward, TupleTableSlot *slot);
-extern bool tuplesort_gettupleslot_mk(Tuplesortstate_mk *state, bool forward, TupleTableSlot *slot);
 
 extern IndexTuple tuplesort_getindextuple(Tuplesortstate *state, bool forward, bool *should_free);
-extern IndexTuple tuplesort_getindextuple_mk(Tuplesortstate_mk *state, bool forward, bool *should_free);
 
 extern bool tuplesort_getdatum(Tuplesortstate *state, bool forward, Datum *val, bool *isNull);
-extern bool tuplesort_getdatum_mk(Tuplesortstate_mk *state, bool forward, Datum *val, bool *isNull);
 
 extern void tuplesort_end(Tuplesortstate *state);
-extern void tuplesort_end_mk(Tuplesortstate_mk *state);
 extern void tuplesort_flush(Tuplesortstate *state);
-extern void tuplesort_flush_mk(Tuplesortstate_mk *state);
 
 extern void tuplesort_finalize_stats(Tuplesortstate *state);
-extern void tuplesort_finalize_stats_mk(Tuplesortstate_mk *state);
 
 extern int	tuplesort_merge_order(long allowedMem);
 
@@ -147,17 +109,11 @@ extern int	tuplesort_merge_order(long allowedMem);
  */
 
 extern void tuplesort_rescan_pos(Tuplesortstate *state, TuplesortPos *pos);
-extern void tuplesort_rescan_pos_mk(Tuplesortstate_mk *state, TuplesortPos_mk *pos);
 extern void tuplesort_rescan(Tuplesortstate *state);
-extern void tuplesort_rescan_mk(Tuplesortstate_mk *state);
 extern void tuplesort_markpos(Tuplesortstate *state);
-extern void tuplesort_markpos_mk(Tuplesortstate_mk *state);
 extern void tuplesort_markpos_pos(Tuplesortstate *state, TuplesortPos *pos);
-extern void tuplesort_markpos_pos_mk(Tuplesortstate_mk *state, TuplesortPos_mk *pos);
 extern void tuplesort_restorepos(Tuplesortstate *state);
-extern void tuplesort_restorepos_mk(Tuplesortstate_mk *state);
 extern void tuplesort_restorepos_pos(Tuplesortstate *state, TuplesortPos *pos);
-extern void tuplesort_restorepos_pos_mk(Tuplesortstate_mk *state, TuplesortPos_mk *pos);
 
 /*
  * tuplesort_set_instrument
@@ -172,10 +128,6 @@ extern void tuplesort_restorepos_pos_mk(Tuplesortstate_mk *state, TuplesortPos_m
  * freed early, disconnect them by calling again with NULL pointers.
  */
 extern void tuplesort_set_instrument(Tuplesortstate *state,
-                         struct Instrumentation    *instrument,
-                         struct StringInfoData     *explainbuf);
-
-extern void tuplesort_set_instrument_mk(Tuplesortstate_mk *state,
                          struct Instrumentation    *instrument,
                          struct StringInfoData     *explainbuf);
 
@@ -196,10 +148,6 @@ extern int32 ApplySortFunction(FmgrInfo *sortFunction, int sortFlags,
 /* Gpmon */
 extern void 
 tuplesort_set_gpmon(Tuplesortstate *state,
-					gpmon_packet_t *gpmon_pkt,
-					int *gpmon_tick);
-extern void 
-tuplesort_set_gpmon_mk(Tuplesortstate_mk *state,
 					gpmon_packet_t *gpmon_pkt,
 					int *gpmon_tick);
